requirements.txt
Flask==2.2.5
flask-cors==3.0.10
numpy==1.24.2
opencv-python-headless==4.8.0.76
Pillow==9.5.0
gunicorn==20.1.0
dlib==19.22.99
